requirements.txt
pandas
psycopg2-binary
python-dotenv
sshtunnel==0.4.0
paramiko==2.12.0
dash
dash-bootstrap-components